+++ packages/terra-content-container/CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Unreleased
 
+* Changed
+  * Removed unnecessary `wdio.conf.js` file
+  * Updated to terra-functional-testing
+
 ## 3.33.0 - (January 5, 2021)
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 3.32.0 - (December 8, 2020)
 
 * Changed