+++ packages/terra-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
+
 ## 1.2.0 - (March 2, 2021)
 
 * Fixed