+++ packages/terra-grid/CHANGELOG.md
@@ -2,11 +2,15 @@
 
 ## Unreleased
 
+* Changed
+  * Removed unnecessary `wdio.conf.js` file
+  * Updated to use terra-functional-testing
+
 ## 6.27.0 - (January 5, 2021)
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 6.26.0 - (December 8, 2020)
 
 * Changed