+++ packages/terra-status-view/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Fixed
   * Fix glyph scrollbar appearing when there is enough room to display.
 
+* Changed
+  * Removed unnecessary `wdio.conf.js` file
+  * Upgrade to terra-functional-testing
+
 ## 4.41.0 - (February 2, 2021)
 
 * Changed
@@ -14,7 +18,7 @@
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 4.39.0 - (December 8, 2020)
 
 * Changed